+++ main.go
@@ -47,6 +47,8 @@ func main() {
 	action_kit_sdk.RegisterAction(extdeployment.NewCheckDeploymentRolloutStatusAction())
 	action_kit_sdk.RegisterAction(extdeployment.NewPodCountCheckAction())
 	action_kit_sdk.RegisterAction(extdeployment.NewPodCountMetricsAction())
+	action_kit_sdk.RegisterAction(extdeployment.NewScaleDeploymentAction())
+	action_kit_sdk.RegisterAction(extstatefulset.NewScaleStatefulSetAction())
 	action_kit_sdk.RegisterAction(extpod.NewDeletePodAction())
 	action_kit_sdk.RegisterAction(extnode.NewNodeCountCheckAction())
 	action_kit_sdk.RegisterAction(extnode.NewDrainNodeAction())