+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 - Attack "Delete Pod" added - :exclamation: Requires new permission `delete` for `pods` resources
 - Attack "Drain node" added - :exclamation: Requires new permission `create` for `pods/eviction` resources and `patch` for `nodes` resources
 - Attack "Taint node" added - :exclamation: Requires new permission `patch` for `nodes` resources
+- Attack "Scale Deployment" added - :exclamation: Requires new permission  `get`, `update` and `patch` for `deployments/scale` resources
+- Attack "Scale StatefulSet" added - :exclamation: Requires new permission `get`, `update` and `patch` for `statefulsets/scale` resources
 - Performance - Add hostnames to kubernetes-deployment during discovery instead of adding it via enrichment rule
 - Added `pprof` endpoints for debugging purposes
 - Memory optimizations